feat(plugin-script): add editor input modes
