+++ packages/plugins/experimental/plugin-script/src/capabilities/react-surface.tsx
@@ -10,6 +10,7 @@ import { ScriptType } from '@dxos/functions';
 import { ScriptCapabilities } from './capabilities';
 import { AutomationPanel, ScriptSettings, ScriptContainer, ScriptSettingsPanel } from '../components';
 import { SCRIPT_PLUGIN } from '../meta';
+import { type ScriptSettingsProps } from '../types';
 
 export default () =>
   contributes(Capabilities.ReactSurface, [
@@ -17,7 +18,10 @@ export default () =>
       id: `${SCRIPT_PLUGIN}/settings`,
       role: 'settings',
       filter: (data): data is any => data.subject === SCRIPT_PLUGIN,
-      component: () => <ScriptSettings settings={{}} />,
+      component: () => {
+        const settings = useCapability(Capabilities.SettingsStore).getStore<ScriptSettingsProps>(SCRIPT_PLUGIN)!.value;
+        return <ScriptSettings settings={settings} />;
+      },
     }),
     createSurface({
       id: `${SCRIPT_PLUGIN}/article`,
@@ -25,7 +29,8 @@ export default () =>
       filter: (data): data is { subject: ScriptType } => data.subject instanceof ScriptType,
       component: ({ data, role }) => {
         const compiler = useCapability(ScriptCapabilities.Compiler);
-        return <ScriptContainer role={role} script={data.subject} env={compiler.environment} />;
+        const settings = useCapability(Capabilities.SettingsStore).getStore<ScriptSettingsProps>(SCRIPT_PLUGIN)!.value;
+        return <ScriptContainer role={role} script={data.subject} settings={settings} env={compiler.environment} />;
       },
     }),
     createSurface({
