+++ packages/plugins/experimental/plugin-script/src/components/ScriptSettings/ScriptSettings.tsx
@@ -6,7 +6,8 @@ import React from 'react';
 
 import { type Credential } from '@dxos/protocols/proto/dxos/halo/credentials';
 import { useClient } from '@dxos/react-client';
-import { Button, useTranslation } from '@dxos/react-ui';
+import { Button, Select, useTranslation } from '@dxos/react-ui';
+import { type EditorInputMode, EditorInputModes } from '@dxos/react-ui-editor';
 import { DeprecatedFormInput } from '@dxos/react-ui-form';
 
 import { SCRIPT_PLUGIN } from '../../meta';
@@ -39,8 +40,31 @@ export const ScriptSettings = ({ settings }: { settings: ScriptSettingsProps }) 
 
   return (
     <>
+      {/* TODO(wittjosiah): Hide outside of dev environments. */}
       <DeprecatedFormInput label={t('authenticate action label')}>
         <Button onClick={handleAuthenticate}>{t('authenticate button label')}</Button>
+      </DeprecatedFormInput>
+
+      <DeprecatedFormInput label={t('editor input mode label')}>
+        <Select.Root
+          value={settings.editorInputMode ?? 'default'}
+          onValueChange={(value) => {
+            settings.editorInputMode = value as EditorInputMode;
+          }}
+        >
+          <Select.TriggerButton placeholder={t('select editor input mode placeholder')} />
+          <Select.Portal>
+            <Select.Content>
+              <Select.Viewport>
+                {EditorInputModes.map((mode) => (
+                  <Select.Option key={mode} value={mode}>
+                    {t(`settings editor input mode ${mode} label`)}
+                  </Select.Option>
+                ))}
+              </Select.Viewport>
+            </Select.Content>
+          </Select.Portal>
+        </Select.Root>
       </DeprecatedFormInput>
     </>
   );
