+++ packages/plugins/experimental/plugin-script/src/translations.ts
@@ -32,6 +32,11 @@ export default [
         'function url label': 'Invocation URL',
         'function binding label': 'Function binding',
         'function binding placeholder': 'BINDING',
+        'editor input mode label': 'Editor input mode',
+        'select editor input mode placeholder': 'Select editor input mode…',
+        'settings editor input mode default label': 'Default',
+        'settings editor input mode vim label': 'Vim',
+        'settings editor input mode vscode label': 'VS Code',
       },
     },
   },
